backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-anthropic
langchain-voyageai
langchain-community
qdrant-client
python-dotenv
pydantic
pydantic-settings
httpx
pypdf
docx2txt
python-docx
markdown
transformers
torch
sentence-transformers
voyageai
motor
langgraph-checkpoint-mongodb
arxiv
minio
